+++ lynglib/src/commonMain/kotlin/net/sergeych/lyng/stdlib_included/observable_lyng.kt
@@ -24,46 +24,46 @@ package lyng.observable
 extern class ChangeRejectionException : Exception
 
 extern class Subscription {
-    fun cancel(): Void
+    extern fun cancel(): Void
 }
 
 extern class Observable<Change> {
-    fun beforeChange(listener: (Change)->Void): Subscription
-    fun onChange(listener: (Change)->Void): Subscription
-    fun changes(): Flow<Change>
+    extern fun beforeChange(listener: (Change)->Void): Subscription
+    extern fun onChange(listener: (Change)->Void): Subscription
+    extern fun changes(): Flow<Change>
 }
 
 extern class ListChange<T>
 
 extern class ListSet<T> : ListChange<T> {
-    val index: Int
-    val oldValue: Object
-    val newValue: Object
+    extern val index: Int
+    extern val oldValue: Object
+    extern val newValue: Object
 }
 
 extern class ListInsert<T> : ListChange<T> {
-    val index: Int
-    val values: List<T>
+    extern val index: Int
+    extern val values: List<T>
 }
 
 extern class ListRemove<T> : ListChange<T> {
-    val index: Int
-    val oldValue: Object
+    extern val index: Int
+    extern val oldValue: Object
 }
 
 extern class ListClear<T> : ListChange<T> {
-    val oldValues: List<T>
+    extern val oldValues: List<T>
 }
 
 extern class ListReorder<T> : ListChange<T> {
-    val oldValues: List<T>
-    val newValues: Object
+    extern val oldValues: List<T>
+    extern val newValues: Object
 }
 
 extern class ObservableList<T> : List<T> {
-    fun beforeChange(listener: (ListChange<T>)->Void): Subscription
-    fun onChange(listener: (ListChange<T>)->Void): Subscription
-    fun changes(): Flow<ListChange<T>>
+    extern fun beforeChange(listener: (ListChange<T>)->Void): Subscription
+    extern fun onChange(listener: (ListChange<T>)->Void): Subscription
+    extern fun changes(): Flow<ListChange<T>>
 }
 
 fun List<T>.observable(): ObservableList<T> {